LayoutTests/imported/w3c/web-platform-tests/css/css-forms/input-text-base-appearance-computed-style-expected.txt


SIBLING

FAIL UA styles of base appearance <input type=text>. assert_equals: background-color expected "rgba(0, 0, 0, 0)" but got "rgb(255, 255, 255)"
